asahi-bless: refactor into a library
